link pages (#25)
* link pages

* fix

* fix
