+++ src/__tests__/components/Loading/Loading.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import "jest";
-import Loading from "../../../components/Loading";
+import Loading from "@/components/Loading";
 import { render } from "@testing-library/react";
 
 describe("Loading", () => {
