+++ src/__tests__/pages/Home/Index.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "jest";
-import Index from "../../../pages/index";
-import { render } from '@testing-library/react';
+import Index from "@/pages/index";
+import { render } from "@testing-library/react";
 
 describe("Index Page", () => {
   it("should render success", () => {
