--- src/components/HeroSection/About/About.module.scss
@@ -1,183 +1,0 @@
-.sectionPricing {
-  padding: 9.6rem 0;
-}
-
-.princingPlan {
-  border-radius: 11px;
-
-  width: 75%;
-}
-
-.planHeading {
-  text-align: center;
-  font-weight: 500;
-  display: block;
-  margin-bottom: 3rem;
-  color: black;
-}
-
-.planName {
-  color: #cf711f;
-  font-weight: 600;
-  font-size: 2rem;
-  text-transform: uppercase;
-  letter-spacing: 0.75;
-  margin-bottom: 3.2rem;
-}
-
-.planPrice {
-  font-size: 6.2rem;
-  font-weight: 600;
-  color: #333;
-  margin-bottom: 1.6rem;
-}
-
-.planPrice span {
-  font-size: 3rem;
-  font-weight: 500;
-  margin-right: 0.8rem;
-}
-
-.planText {
-  font-size: 1.6rem;
-  line-height: 1.6;
-  color: #6f6f6f;
-}
-
-.planSingUp {
-  text-align: center;
-  margin-top: 4.8rem;
-}
-
-.planDetails {
-  font-size: 1.6rem;
-  line-height: 1.6;
-  text-align: center;
-}
-
-.line {
-  display: block;
-  color: gray;
-  margin-end: auto;
-  margin-start: auto;
-  border: 1px inset;
-  overflow: hidden;
-}
-
-.container {
-  max-width: 120rem;
-  padding: 0 3.2rem;
-  margin: 0 auto;
-  text-align: center;
-  width: 40%;
-}
-
-.btnContainer {
-  max-width: 120rem;
-  padding: 0 3.2rem;
-  margin: 0 auto;
-  align-items: center;
-  justify-content: center;
-  display: flex;
-}
-
-.homeSwitchingTo_cta {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
-}
-
-.btn {
-  border: none;
-  outline: none;
-  background-color: red;
-  color: #fff;
-  font-weight: 600;
-  font-size: 16px;
-  line-height: 150%;
-  letter-spacing: 0.75px;
-  padding: 16px 18px;
-  border-radius: 2rem;
-  margin-right: 1rem;
-}
-
-.productTour {
-  font-size: 14px;
-  font-weight: 500;
-  line-height: 157%;
-  letter-spacing: 0.25px;
-  color: #727491;
-  flex: 1;
-  color: red;
-}
-
-.gridContainer {
-  max-width: 120rem;
-  padding: 0 3.2rem;
-  margin: 0 auto;
-  display: grid;
-  column-gap: 6.4rem;
-  row-gap: 9.6rem;
-  grid-template-columns: repeat(3, 1fr);
-}
-
-.planHeading {
-  font-size: 2rem;
-  margin-bottom: 1rem;
-  color: black;
-}
-
-.planHeadingSectiondary {
-  font-size: 1rem;
-  margin-bottom: 1rem;
-  color: black;
-}
-
-.frontend:hover {
-  transform: translateY(-1.2rem);
-  box-shadow: 0 3.2rem 6.4rem rgba(0, 0, 0, 0.06);
-}
-
-.center-text {
-  text-align: center;
-}
-
-.btn--form:hover {
-  background-color: #fff;
-  color: #555;
-  align-self: end;
-}
-
-.price {
-  box-shadow: 0 2.4rem 4.8rem rgba(0, 0, 0, 0.075);
-  border-radius: 11px;
-  overflow: hidden;
-  transition: all 0.4s;
-}
-
-.priceContent {
-  padding: 3.2rem 4.8rem 4.8rem 4.8rem;
-}
-
-.priceTitle {
-  font-size: 2.4rem;
-  color: #333;
-  font-weight: 600;
-  margin-bottom: 1rem;
-  display: block;
-}
-
-.priceAttributes {
-  list-style: none;
-  display: flex;
-  text-align: left;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.priceAttribute {
-  font-size: 15px;
-  display: flex;
-  align-items: center;
-  gap: 1.6rem;
-}
