+++ src/components/HeroSection/AccessInventory/index.tsx
