+++ src/components/HeroSection/SyncInventory/index.tsx
