+++ src/pages/index.tsx
@@ -1,71 +1,34 @@
-import type { NextPage } from "next";
-import Avatar from "@mui/material/Avatar";
-import Button from "@mui/material/Button";
-import TextField from "@mui/material/TextField";
-import FormControlLabel from "@mui/material/FormControlLabel";
-import Checkbox from "@mui/material/Checkbox";
-import Box from "@mui/material/Box";
-import Typography from "@mui/material/Typography";
 import Container from "@mui/material/Container";
+import OrganizeInventory from "@/components/HeroSection/OrganizeInventory";
+import AccessInventory from "@/components/HeroSection/AccessInventory";
+import GetReporting from "@/components/HeroSection/GetReporting";
+import SyncInventory from "@/components/HeroSection/SyncInventory";
+import BusinessPartner from "@/components/HeroSection/BussinessPartner";
+import Download from "@/components/HeroSection/Download";
+import Pricing from "@/components/HeroSection/Pricing";
+import GetStart from "@/components/HeroSection/GetStart";
+import About from "@/components/HeroSection/About";
+import Header from "@/components/Header";
+import Footer from "@/components/Footer";
 
-// https://mui.com/zh/material-ui/material-icons/
-import LockOutlinedIcon from "@mui/icons-material/LockOutlined";
-
-const Index: NextPage = () => {
-  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    const data = new FormData(event.currentTarget);
-    console.log(data);
-  };
-
+function index() {
   return (
-    <Container component="main" maxWidth="xs">
-      <Box
-        sx={{
-          marginTop: 8,
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-        }}
-      >
-        <Avatar sx={{ m: 1, bgcolor: "secondary.main" }}>
-          <LockOutlinedIcon />
-        </Avatar>
-        <Typography component="h1" variant="h5">
-          Sign in
-        </Typography>
-        <Box component="form" onSubmit={handleSubmit} noValidate sx={{ mt: 1 }}>
-          <TextField
-            margin="normal"
-            required
-            fullWidth
-            id="email"
-            label="Email Address"
-            name="email"
-            autoComplete="email"
-            autoFocus
-          />
-          <TextField
-            margin="normal"
-            required
-            fullWidth
-            name="password"
-            label="Password"
-            type="password"
-            id="password"
-            autoComplete="current-password"
-          />
-          <FormControlLabel
-            control={<Checkbox value="remember" color="primary" />}
-            label="Remember me"
-          />
-          <Button type="submit" fullWidth variant="contained" sx={{ mt: 3, mb: 2 }}>
-            Sign In
-          </Button>
-        </Box>
-      </Box>
-    </Container>
+    <>
+      <Header />
+      <Container>
+        <OrganizeInventory />
+        <BusinessPartner />
+        <AccessInventory />
+        <GetReporting />
+        <SyncInventory />
+        <Download />
+        <Pricing />
+        <About />
+        <GetStart />
+      </Container>
+      <Footer />
+    </>
   );
-};
+}
 
-export default Index;
+export default index;
